+++ src/router/components.ts
@@ -19,7 +19,7 @@ export function Routes(props: {
 	return nest(() => {
 		const match = watched.match;
 		if (match) {
-			return () => inject([ROUTER, new ChildRouter(router, match.path, () => watched.rest)], () => {
+			return () => inject(ROUTER, new ChildRouter(router, match.path, () => watched.rest), () => {
 				return match.route.content({ params: match.params });
 			});
 		}